requirements.txt
Flask==3.0.0
flask-mysqldb==2.0.0